// src/app/models/ActeTraitement.ts
export class ActeTraitement{
    idactetrait?: String;
    ref_tacheBPU?: String;
    type_prestation?: String;
    type_element?: String;
    quantite?: String;
	 date_reception?: String;
	 date_livraison?: String;
	 date_validation?: String;
	 affectation?: String;
	duree?: String;
	 commentaire?: String;
	 motif?: String;
	 statut_facturation?: String;
	date_reprise?: String;
    reprise_facturable?: String;
}